RequirementMachine: Split off RuleBuilder into a new RequirementLowering.cpp file
